+++ resources/firewalld_helpers.rb
@@ -7,7 +7,7 @@ property :version,
          String,
          default: '',
          description: 'see version attribute of helper tag in firewalld.helper(5).'
-property :name,
+property :short,
          String,
          name_property: true,
          description: 'see short tag in firewalld.helper(5).'
@@ -15,36 +15,37 @@ property :description,
          String,
          description: 'see description tag in firewalld.helper(5).'
 property :family,
-         [String, Array],
-         equal_to: ['ipv4', 'ipv6', %w(ipv4 ipv6)],
+         String,
+         equal_to: %w(ipv4 ipv6),
          default: 'ipv4',
          description: 'see family tag in firewalld.helper(5).'
 property :nf_module,
          String,
          description: 'see module tag in firewalld.helper(5).'
 property :ports,
-         Array,
+         [Array, String],
          default: [],
-         description: 'array of port and protocol pairs. See port tag in firewalld.helper(5).'
+         description: 'array of port and protocol pairs. See port tag in firewalld.helper(5).',
+         coerce: proc { |o| Array(o) }
 
 load_current_value do |new_resource|
-  sysbus = DBus.system_bus
-  firewalld_service = sysbus['org.fedoraproject.FirewallD1']
+  dbus = DBus.system_bus
+  firewalld_service = dbus['org.fedoraproject.FirewallD1']
   firewalld_object = firewalld_service['/org/fedoraproject/FirewallD1/config']
-  interface = firewalld_object['org.fedoraproject.FirewallD1.config']
-  begin
-    helper_path = interface.getHelperByName(new_resource.name)
+  fw_config = firewalld_object['org.fedoraproject.FirewallD1.config']
+  if fw_config.getHelperNames.include?(new_resource.short)
+    helper_path = fw_config.getHelperByName(new_resource.short)
     object = firewalld_service[helper_path]
-    interface = object['org.fedoraproject.FirewallD1.config.helper']
-    settings = interface.getSettings
+    config_helper = object['org.fedoraproject.FirewallD1.config.helper']
+    settings = config_helper.getSettings
     version settings[0]
-    # name settings[1]
+    # short settings[1]
     description settings[2]
     family settings[3]
     nf_module settings[4]
     ports settings[5]
-  rescue DBus::Error
-    Chef::Log.info "Helper #{new_resource.name} does not exist. Will be created."
+  else
+    Chef::Log.info "Helper #{new_resource.short} does not exist. Will be created."
   end
 end
 
@@ -52,29 +53,44 @@ action :update do
   dbus = DBus.system_bus
   fw = firewalld_interface(dbus)
   fw_config = config_interface(dbus)
+  helper_names = fw_config.getHelperNames
   reload = false
-  begin
-    helper_path = interface.getHelperByName(new_resource.name)
-    helper = helper_interface(dbus, helper_path)
-
-    reload = false
-    values = properties.map do |property|
-      new_resource.send(property)
-    end
-    helper.update(values)
-  rescue
+  if !helper_names.include?(new_resource.short)
     values = [
       new_resource.version,
-      new_resource.name,
-      new_resource.description || default_description(new_resource),
+      new_resource.short,
+      default_description(new_resource),
       new_resource.family,
       new_resource.nf_module,
-      new_resource.ports,
+      new_resource.ports.map { |e| e.split('/') },
     ]
-    converge_by "Add Helper #{new_resource.name}" do
-      fw_config.addHelper(new_resource.name, values)
+    converge_by "Add Helper #{new_resource.short}" do
+      fw_config.addHelper(new_resource.short, values)
     end
     reload = true
+  else
+    helper_path = fw_config.getHelperByName(new_resource.short)
+    helper = helper_interface(dbus, helper_path)
+    converge_if_changed :version do
+      helper.setVersion new_resource.version
+      reload = true
+    end
+    converge_if_changed :description do
+      helper.setDescription default_description(new_resource)
+      reload = true
+    end
+    converge_if_changed :family do
+      helper.setFamily new_resource.family
+      reload = true
+    end
+    converge_if_changed :nf_module do
+      helper.setModule new_resource.nf_module
+      reload = true
+    end
+    converge_if_changed :ports do
+      helper.setPorts new_resource.ports.map { |e| e.split('/') }
+      reload = true
+    end
   end
 
   if reload